+++ crates/l2/sequencer/block_producer.rs
@@ -116,20 +116,21 @@ impl BlockProducer {
         let chain_config = store.get_chain_config()?;
         validate_block(&block, &head_header, &chain_config)?;
 
+        let account_updates = payload_build_result.account_updates;
+
         let execution_result = BlockExecutionResult {
-            account_updates: payload_build_result.account_updates,
             receipts: payload_build_result.receipts,
             requests: Vec::new(),
         };
 
         blockchain
-            .store_block(&block, execution_result.clone())
+            .store_block(&block, execution_result.clone(), &account_updates)
             .await?;
         info!("Stored new block {:x}", block.hash());
         // WARN: We're not storing the payload into the Store because there's no use to it by the L2 for now.
 
         // Cache execution result
-        execution_cache.push(block.hash(), execution_result.account_updates)?;
+        execution_cache.push(block.hash(), account_updates)?;
 
         // Make the new head be part of the canonical chain
         apply_fork_choice(&store, block.hash(), block.hash(), block.hash()).await?;