+++ crates/l2/sequencer/l1_committer.rs
@@ -150,9 +150,13 @@ impl Committer {
                 warn!(
                             "Could not find execution cache result for block {block_number}, falling back to re-execution"
                         );
-                Evm::default(self.store.clone(), block_to_commit.header.parent_hash)
-                    .execute_block(&block_to_commit)
-                    .map(|result| result.account_updates)?
+                let mut vm = Evm::default(self.store.clone(), block_to_commit.header.parent_hash);
+                vm.execute_block(&block_to_commit)?;
+                let fork = self
+                    .store
+                    .get_chain_config()?
+                    .fork(block_to_commit.header.timestamp);
+                vm.get_state_transitions(fork)?
             }
         };