+++ crates/vm/backends/revm/mod.rs
@@ -97,13 +97,7 @@ impl REVM {
             }
         }
 
-        let account_updates = Self::get_state_transitions(state);
-
-        Ok(BlockExecutionResult {
-            receipts,
-            requests,
-            account_updates,
-        })
+        Ok(BlockExecutionResult { receipts, requests })
     }
 
     pub fn execute_tx(